% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLC.R
\name{MLC}
\alias{MLC}
\title{MLC}
\usage{
MLC(x, y, ridge = 1e-09)
}
\arguments{
\item{x}{feature vector}

\item{y}{response vector}
}
\value{
An object of class \code{\link[tabularMLC]{MLC.model-class}} parameters used for the model
}
\description{
Function to create classifier class with dataset
}
