% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLC.R
\name{predict,MLC.model-method}
\alias{predict,MLC.model-method}
\title{Predict function for class MLC.model}
\usage{
\S4method{predict}{MLC.model}(object, x, likelihood = FALSE)
}
\description{
Predict function for class MLC.model
}
