% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLC.R
\name{predict.MLC}
\alias{predict.MLC}
\title{predict.MLC}
\usage{
predict.MLC(model, x, likelihood = FALSE)
}
\arguments{
\item{model}{\code{\link[tabularMLC]{MLC.model-class}} model class to use for prediction}

\item{x}{feature vector to predict}
}
\value{
a factor vector with the predicted value
}
\description{
Function to create classifier class with dataset
}
