% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularMLC.R
\docType{package}
\name{tabularMLC-package}
\alias{tabularMLC}
\alias{tabularMLC-package}
\title{Tabular maximum likelihood classifier}
\description{
Maximum likelihood is a common classifier used for land use classification.
It calculates the likelihood of an object to belong to each class based on
an expected distribution and a metric of distance.
}
\details{
The most common implementation, like in this package, will assume normal
distributed variables within classes, and calculate the distance, based
on Mahalanobis distance.
}
\references{
Mather, P. M. (1985). Remote sensing letters: A computationally efficient 
maximum-likelihood classifier employing prior probabilities for remotely-sensed 
data. International Journal of Remote Sensing, 6(2), 369–376. 
\doi{10.1080/01431168508948456}

Imports
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/caiohamamura/tabularMLC}
  \item Report bugs at \url{https://github.com/caiohamamura/tabularMLC/issues}
}

}
\author{
\strong{Maintainer}: Caio Hamamura \email{caiohamamura@gmail.com} (\href{https://orcid.org/0000-0001-6149-5885}{ORCID})

}
